refactor: add new icons
