+++ src/componenets/NewGame.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
-import { ReactComponent as Mine } from '../assets/mine.svg'
+import { ReactComponent as PlayingIcon } from '../assets/playing.svg'
 
 export default function NewGame() {
 
   return (
     <div className="header__new-game">
-      <Mine />
+      <PlayingIcon />
     </div>
   )
 }
