+++ src/componenets/RecordGame.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import { ReactComponent as PlayIcon } from '../assets/play.svg'
+
+export default function RecordGame() {
+
+  return (
+    <div className="header__record-game">
+      <PlayIcon />
+    </div>
+  )
+}
