+++ src/componenets/TopScores.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
-import { ReactComponent as Mine } from '../assets/mine.svg'
+import { ReactComponent as TopScoresIcon } from '../assets/podium.svg'
 
 export default function TopScores() {
 
   return (
     <div className="header__top-scores">
-      <Mine />
+      <TopScoresIcon />
     </div>
   )
 }
